PR comment fix - Remove navigation graphs
